Renamet Mellom1Og2-vilkår og justert dato-validering
