+++ src/frontend/komponenter/Fagsak/Vilkårsvurdering/GeneriskVilkår/Vilkår/Mellom1Og2EllerAdoptert/Mellom1Og2EllerAdoptert.tsx
@@ -3,18 +3,18 @@ import React from 'react';
 import type { IVilkårSkjemaBaseProps } from '../../VilkårSkjema';
 import { VilkårSkjema } from '../../VilkårSkjema';
 import { useVilkårSkjema } from '../../VilkårSkjemaContext';
-import { useMellom1Og2EllerAdopsjon } from './Mellom1Og2EllerAdopsjonContext';
+import { useMellom1Og2EllerAdoptert } from './Mellom1Og2EllerAdoptertContext';
 
 type Mellom1Og2EllerAdopsjonProps = IVilkårSkjemaBaseProps;
 
-export const Mellom1Og2EllerAdopsjon: React.FC<Mellom1Og2EllerAdopsjonProps> = ({
+export const Mellom1Og2EllerAdoptert: React.FC<Mellom1Og2EllerAdopsjonProps> = ({
     vilkårResultat,
     vilkårFraConfig,
     toggleForm,
     person,
     lesevisning,
 }: Mellom1Og2EllerAdopsjonProps) => {
-    const { felter } = useMellom1Og2EllerAdopsjon(vilkårResultat, person);
+    const { felter } = useMellom1Og2EllerAdoptert(vilkårResultat, person);
     const vilkårSkjemaContext = useVilkårSkjema(vilkårResultat, felter, person, toggleForm);
     return (
         <VilkårSkjema
