+++ src/frontend/komponenter/Fagsak/Vilkårsvurdering/GeneriskVilkår/Vilkår/Mellom1Og2EllerAdoptert/Mellom1Og2EllerAdoptertValidering.ts
