+++ src/frontend/komponenter/Fagsak/Vilkårsvurdering/GeneriskVilkår/VilkårTabellRad.tsx
@@ -22,7 +22,7 @@ import { BorMedSøker } from './Vilkår/BorMedSøker/BorMedSøker';
 import { BosattIRiket } from './Vilkår/BosattIRiket/BosattIRiket';
 import { Medlemskap } from './Vilkår/Medlemskap/Medlemskap';
 import { MedlemskapAnnenForelder } from './Vilkår/MedlemskapAnnenForelder/MedlemskapAnnenForelder';
-import { Mellom1Og2EllerAdopsjon } from './Vilkår/Mellom1Og2EllerAdopsjon/Mellom1Og2EllerAdopsjon';
+import { Mellom1Og2EllerAdoptert } from './Vilkår/Mellom1Og2EllerAdoptert/Mellom1Og2EllerAdoptert';
 import { vilkårFeilmeldingId } from './VilkårTabell';
 
 interface IProps {
@@ -135,7 +135,7 @@ const VilkårTabellRad: React.FC<IProps> = ({ person, vilkårFraConfig, vilkårR
                 );
             case VilkårType.MELLOM_1_OG_2_ELLER_ADOPTERT:
                 return (
-                    <Mellom1Og2EllerAdopsjon
+                    <Mellom1Og2EllerAdoptert
                         vilkårResultat={vilkårResultat}
                         vilkårFraConfig={vilkårFraConfig}
                         toggleForm={toggleForm}
